Use 32-bit arithmetic for entanglement.

+++ nikon/coolpix_a900/sub.h
@@ -7,7 +7,7 @@
 namespace ls_sec {
 
 uint64_t entangle(const std::vector<uint64_t> &elements);
-std::pair<uint64_t, uint64_t> sub_1480(uint64_t a, uint64_t b);
+std::pair<uint32_t, uint32_t> sub_1480(uint32_t a, uint32_t b);
 uint32_t sub_1A40(uint32_t src);
 
 } // namespace ls_sec
