+++ nikon/coolpix_a900/sub_data.h
@@ -3,6 +3,6 @@
 #include <array>
 #include <cstdint>
 
-extern std::array<uint64_t, 1048> DATA_2A00;
+extern std::array<uint32_t, 1048> DATA_2A00;
 
 extern std::array<uint64_t, 8> DATA_4AC0;
